requirements.txt
# CollegeGPT Dependencies
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1

# LLM & Embeddings
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
langchain-text-splitters==0.3.4
google-generativeai==0.8.4

# Vector Database
faiss-cpu==1.9.0.post1

# PDF Processing
pymupdf==1.25.3

# Interface
streamlit==1.41.1
